Changed from IconList to IconMap
DON'T ACCEPT THIS INTO THE PULL REQUEST. A map feels more flexible and
easier to use than a list. It also removes some of the out of bounds
problems we were having. Right now hte program should work the same as
before but with a map instead. The try-catch is not tested yet as ive
been trying to get the setIcon function to work first. Pls look at it
say what you think.