Save analysis area and don't override save data
If an analaysis is done within an interval or a bounding box they are saved and can be accessed by the context menu of the analysis.

when an analysis is done it will now check if another analysis with the same name already is saved and add a number to the end if that's the case to prevent overwriting.