Remove project (#4)
* Removed the crash when you close the project

Just disabled the buttons.

It does however still crash if you rightclick the folder with the current video.

* No longer crashes when removing a folder

If the current vid_proj is removed the video is stopped and all video buttons turned of. The video_widget is still however showing a picture.

* Frame widget now closes when the video is removed

* Fixed crash when loading a new project

Crash happened when opening a new project while a video was not on the first frame. Videoplayer now checks if a video is loaded before processing a frame.

* Cleaned up a little

* Cleaned up according to pull request comments