Queue widget is now a dockable (#9)
* Fixed some smaller things related to the video buttons

Added a crosshair cursor when drawing an analysis area. Made the simple analysis button toggleable instead. And trying to correct the view menu with the analysis queue.

* Fixed the confirm dialog when choosing analysis area and fixed the show/hide drawing

also fixed a bug that could crash the program in the queuedialog.

* Changing queue_wgt to a dockable

Not done yet, still have some show/hides to fix

* Small cleanup

* Show/hide now works

Dock widget now show/hide correctly when analyses start and ends

* Changes from github

Widget only closes if it is floating and added a nother repaint