Created a drawing widget
Added a new dockable, drawingwidget.
This widget has a list of all the frames that have drawings on them in the current video project.
By clicking a frame the video will jump to it.
By clicking a drawing the video will jump to it and the drawing will be set to current drawing.
Context menu with rename and remove exist. By removing a drawing it gets deleted and removeing a frame will remove that frame from the list and delete all drawings on the frame.