Save and rename drawings
Now works on 4 draw types, rectangle, circle, line & arrow
Saving works
Rename works on the drawings, removes from the frames since they are constant.
If the last drawing on a frame are remove the frame is removed as well.
removed the undo/redo functionality (buttons are currently still there)
The drawings are now Shapeitems onstead of treeitems.