Update text and drag/drop
The context menu for text have an option to change the text.
Enabled drag for the drawings and drop only in frame items.

TODO
not drag drawing to another frame
change order in vector
Fix crash when u have more than 3 drawings and restart the program 3 or more times, yeah