Finally added remove on the items in the project widget
The items in the project widget are now removable. The function iterates recursively and deletes the item and all of its children.
Basic analyses now holds their IDs to make it simple to get them from the list of basic analyses.

Also, when adding a new video to the tree if a folder is selected, the video will now get added in that folder.