Color tool (#31)
* Created a drawing widget

Added a new dockable, drawingwidget.
This widget has a list of all the frames that have drawings on them in the current video project.
By clicking a frame the video will jump to it.
By clicking a drawing the video will jump to it and the drawing will be set to current drawing.
Context menu with rename and remove exist. By removing a drawing it gets deleted and removeing a frame will remove that frame from the list and delete all drawings on the frame.

* Save and rename drawings

Now works on 4 draw types, rectangle, circle, line & arrow
Saving works
Rename works on the drawings, removes from the frames since they are constant.
If the last drawing on a frame are remove the frame is removed as well.
removed the undo/redo functionality (buttons are currently still there)
The drawings are now Shapeitems onstead of treeitems.

* Update text and drag/drop

The context menu for text have an option to change the text.
Enabled drag for the drawings and drop only in frame items.

TODO
not drag drawing to another frame
change order in vector
Fix crash when u have more than 3 drawings and restart the program 3 or more times, yeah

* Change edit shape

Changed how edit shape works
Added a delete shortcut for deleting the currently marked object in the drawing widget

* Fixed some crashes

The shape's names now updates correctly, which mean it no longer should say "unknown shape" after changing video.
Fixed a crash when having too many drawings (or something like that). It now iterate through the shapes in a vecor correctly.
Added a clear on the drawing widget when closing a video project.
Started changing how adding a new text works, not done.

* Remove the text rework

wait with it till the connect tool, since need to connect more signals from the drawing widget. Gonna need to connect signals back froms drawing widget to the overlay to show current item.

* Fix scroll wheel crash

Fixed crash when using scroll wheel while  no current drawing was selected.

* Updated edit shape

The point to edit is now selected upon the first mouse click and not based on the closeds point. To edit the other point just release the mouse and click close to the other point.

* Added change text size

You can change the text size by either drawing it lika a normal drawing or by using the scroll wheel. This means that there is currently no way to change the thickness of a text drawing.

* Update the draw end of the text

Updates the draw end coordinate to match the size of the text.

* Added the pen item and removed the ability to select items from the frame

Drawings can now only be selected from the drawing widget. This since it wasn't necessary and removed some problems.
Also added the pen item so it can be used in the drawing widget.

* Changed current drawing selection

Current drawing selection is now a dashed line around the drawing instead of inverting the color. The selection now neither is saved or shows up on the bookmarks.

* Text on frame and in widget are now mirrored

When updating the text in the widget the text on the frame are now updated. Works both with the rename option and doubleclicking on the item.

by removing the signal blocking in the add item to frame the save item data might not be needed.

* Cleaned up the code

fixed a virtual function in ShapeItem to get cleaner switches in drawingwidget

* Moved names to the cildren of shapes

It's now the children how have the names and they are stores as a Qstring.

All things renaming the drawings in the drawing widget should work.

* Added destructors

Destructors
Cleaned the current drawing rect after the item is deleted.

* Current item

Current item are now set in the drawing widget when loading a new overlay

* Added some comments

Comments in drawingwidget

* Fixed a crash with show/hide drawings

* Renamed rename and removed undo/redo

* Oops

Remove delete which shouldn't have been there.

* Bug fixes

Fix save prompt and a crash with delete action and deselect drawing in the widget by clicking the frame item.

* Fix a merge miss

* Drawings are movable and savable

Drawings with the pen are now saved as a vector of points instead of pairs of points.

TODO the old shell when they were pairs are still here. remove? removeing means old pen drawings will not appear

* Added an indicator for current color

Current color is now showed next to the color chooser action. It updates when the current color changes.

* Update drawingtoolbar.h