Analysissettings will now hold the settings for analyses
Settings chosen in the advance analysis window will be saved in the analysis settings. This will make them retrievable later and the specific settings for an analysis can be displayed. They can also be saved between analyses.
Project widget now holds a analysissettings will the currently chosen settings which can be modified in the settings on the analysis menu.