Remove, clean (#44)
* Finally added remove on the items in the project widget

The items in the project widget are now removable. The function iterates recursively and deletes the item and all of its children.
Basic analyses now holds their IDs to make it simple to get them from the list of basic analyses.

Also, when adding a new video to the tree if a folder is selected, the video will now get added in that folder.

* Added more deletes

Make sure the tree item and the data is deleted to avoid memory leaks.

* Cleaned both code and memory

Todo look at and maybe change analysis method with the proxy and the analysis for fixing the basic analysis destructor

* Added some more deletes

Added some more deletes to clean up all memory when closing a project

TODO clean overlay

* Fix deletes when removing projects

* Proxy is created as a pointer and added some deletes

Cleaning up the basic analysises a little. Analysis slider won't now load the entire analysis when it wants the analysis interval.

* Cleaned up some old code

Remove no longer needed functions and updates some comments in detectionbox and analysisproxy.

* Lots of debug messages

Trying to get the destructors for all analysis to work as wanted.

TODO Currently you can only add an analysis or delete an analysis correctly but not both at the same time.

* Memory leaking analyses

Should save, load and remove analyses without too many memory leaks. It does currently however save the temp map in the analyses, so that needs to be fixed.

TODO might be able to move the clean intervals back to basic analysis if they are even needed.

* Fixing some bugs after the merge with development

Added a delete for drawingtag.
Don't save frame with no drawings.
Added some checks for if tree_item_clicked was called with an empty items.
Fixed a problem with  showing tags on the slider.
And fixed the save path for the analysis.

TODO Save/load analyses.

* Fixing some bugs

* Testing

* Fixed bug

* Clear ana interval when clicking another video

* remove analysis from map when deleted

* Soo many comments (less)

* Refractored remove tree item

And removed some comments.

* Oops