Big changes on on_tree_item_clicked
No longer recursive. Fixed it to work with the new Tag_frame_item nad how tags should work with all the frames as subitems. Tags have a checkbox which indicate the current tag. It's this tag that new tags will be created in. The current tag is the same while clicking around in a video project.

analyses are now saved in the framewidget till a new one is loaded. This will save a lot of loading time if there's only one analysis.