Some smaller fixed to bookmarks
Added the current time in the description.
Changed the create and edit bookmark windows. It no longer uses the custom dialog.
Added Ctrl + enter as a shortcut to close the edit description windows.
The windows now have focus when opened. This make it easier to write the description as you don't have to select the dialog first.
Remove a way for bookmarks ro be created even if cancel was clicked.
If the description is empty if won't show the "Description:" in the tool tip and probably won't write it in the report (not tested yet)