Perform analysis dialog 2 (#49)
* Added destructor

* Cleaning & small fixes

The ok button in the perform analysis dialog is greyed out until a video is selected.

* Changed detect shadows to a checkbox

* Puffed up the analysis menu and cleaned mainwindow

With some fancy new icons

* Added settings to the analysis menu

Settings are added to the analysis menu. It's the same window as the advanced analaysis but without the video list and the analysis settings are saved and used as the default values when doing a quick analysis or when opening the advanced analysis.

* Added reset in settings

* Analysissettings will now hold the settings for analyses

Settings chosen in the advance analysis window will be saved in the analysis settings. This will make them retrievable later and the specific settings for an analysis can be displayed. They can also be saved between analyses.
Project widget now holds a analysissettings will the currently chosen settings which can be modified in the settings on the analysis menu.

* Highlight analyses when done and select current video in analysis dialog

The current video is now selected in the analysis dialog.
Highlight an analysis in the project tree when it's done analyzing. The highlighting are a NEW string and setting the background color to yellow. The highlight disappear when clicked.

* Added a new dockable widget for showing the settings for a analysis

In the context menu for an analysis there's a new option, "Show settings" which will show the variables used for that analysis.

TODO
Add interval, boundingbox, ana interval, video, name.

* Fixed so interval and b box works again and added them to the analysis settings widget

* Changed from feedback

* Fix Save load analysis settings

And filled out the settings with the real settings.
Fixed a crash so you couldn't click an analysis items.

* Bounding box now saves in the analysis and therefore shows up from the framewidget

* Added checkboxes on the settings on the context menu

* Cleaned the code

Removed comments, added comments

* Changed the name of the analysis dialog window