Tags and state (#50)
* the frames in a tag now shows under the tag items

Still not all functionality.

* New tag class and remade the add and remove tags to work for TagFrameItem

* New tag with frames

Todo, crashes when loading and clicking a tag of frame

* Debugging

* Some comments

* Checkbox for tags

the box checked means the tags is the current tag. Tagged frames will go to the current tag even if the tag isnt checked.

Deselect the tag when change tags or changing video or clicking the tag again. TODO currently the tag check/unchecks on tagframe items

* Big changes on on_tree_item_clicked

No longer recursive. Fixed it to work with the new Tag_frame_item nad how tags should work with all the frames as subitems. Tags have a checkbox which indicate the current tag. It's this tag that new tags will be created in. The current tag is the same while clicking around in a video project.

analyses are now saved in the framewidget till a new one is loaded. This will save a lot of loading time if there's only one analysis.

* Fixed having multiple tags under same video

* Fixed next/prev poi btn

Jump to next/prev frame in the tag or poi in analysis

* Start remake tags again

Tags should save the video state so the y can save the zoom rect and scale factor

* Revert "Start remake tags again"

This reverts commit e632560081f09ec96b82d476fd1cd1cfe875845e.

* Video state remembers the frame

Fixed some problems with loading a video on a certain frame which were needed for tags.

* Tags now hold a map of TagFrame items

These TagFrame items holds the video state from when they were tagged.

should work the same as they did before the change to a map with the new obejct instead of a set of ints.

* Cleaned code

Removed/added comments

* Oops

* changed some function names

and fixed some other comments on the pull .request