Bookmarks fixes (#58)
* Some smaller fixed to bookmarks

Added the current time in the description.
Changed the create and edit bookmark windows. It no longer uses the custom dialog.
Added Ctrl + enter as a shortcut to close the edit description windows.
The windows now have focus when opened. This make it easier to write the description as you don't have to select the dialog first.
Remove a way for bookmarks ro be created even if cancel was clicked.
If the description is empty if won't show the "Description:" in the tool tip and probably won't write it in the report (not tested yet)

* cleaned some code

removed the include for custom dialog and using std namespace which required some fixing

* Changed some variable names and starting to work on copy constructor for copying bookmarks

* Each bookmark is unique should only be in one container

Copying a bookmark will now create a new bookmark and not just a new pointer to the same bookmark

* Container now saved as one pair instead of vector of pairs

Each bookmark can only have one container

* Bookmarks now save the state when they are created

* Bookmarks are now removed as they should

To prevent them from leaking too much memory.

* Bookmarks are inserted on correct index when moved

Containers are not inserted correctly.
Still using the old user written code.

* Containers are inserted on the correct index and clicking the bookmarks will jump to correct state

* Bookmarks are removed when deleted, added context menu with new category

Clean code and added comments

TODO remove bookmarks from folders

* Move add new folder to bookmarklist

Simplify the destructor in bookmarklist