Changed how the program will remove analysis
When removing an analysis the path to the analysis is saved in a list and the file in the folder in untouched. Later when the project is saved the list of "removed" project is removed from the folder. If the project is not saved the list is just cleared so the file are still in the folder.