videoproject have its own state
Video project have its own state. This will give the video a state that it will jump back to when clicking the video again. This state is only updated if the video is the last thing clicked in the project tree, jumping to a tag will not update it

TODO save color corrections as well